Refactor  convert_header/CCP4/convert.py to headers.py
