+++ pyworkflow/em/packages/ccp4/protocol_refmac.py
@@ -31,7 +31,7 @@ import pyworkflow.protocol.constants as const
 from constants import CCP4VERSION, CCP4VERSIONFILENAME
 from pyworkflow import VERSION_1_2
 from pyworkflow.em import PdbFile
-from pyworkflow.em.headers.ccp4header import (
+from pyworkflow.em.headers import (
     adaptFileToCCP4, START, Ccp4Header)
 from pyworkflow.em.packages.ccp4.convert import (runCCP4Program, getProgram)
 from pyworkflow.em.packages.ccp4.refmac_template_map2mtz import \
